Replace size_t with int32_t

+++ drivers/stack_struct/stack_private.h
@@ -13,8 +13,8 @@
 typedef struct
 {
 	void * array;
-	size_t stack_pointer;
-	size_t size;
+	int32_t stack_pointer;
+	int32_t size;
 } stack_t;
